% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Calc_Anisotropic_Mesh.R
\name{Calc_Anisotropic_Mesh}
\alias{Calc_Anisotropic_Mesh}
\title{Make mesh for distances among points}
\usage{
Calc_Anisotropic_Mesh(loc_x, Method = "Mesh", anisotropic_mesh = NULL,
  refine = FALSE, ...)
}
\arguments{
\item{loc_x}{location (eastings and northings in kilometers, UTM) for each sample or knot}

\item{Method}{spatial method determines ("Mesh" and "Grid" give}

\item{anisotropic_mesh}{OPTIONAL, anisotropic mesh (if missing, its recalculated from loc_x)}

\item{refine}{OPTIONAL, specify whether to add additional points (beyond loc_x and minimal boundary knots)}
}
\value{
Tagged list containing distance metrics
}
\description{
\code{Calc_Anistropic_Mesh} builds a tagged list representing distances for isotropic or geometric anisotropic triangulated mesh
}

