% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc_Kmeans.R
\name{Calc_Kmeans}
\alias{Calc_Kmeans}
\title{Calculate location for knots approximating spatial variation}
\usage{
Calc_Kmeans(n_x, loc_orig, nstart = 100, randomseed = NULL,
  iter.max = 1000, DirPath = NULL, Save_Results = TRUE)
}
\arguments{
\item{n_x}{the number of knots to select}

\item{loc_orig}{a matrix with two columns where each row gives the 2-dimensional coordinates to be approximated}

\item{nstart}{the number of times that the k-means algorithm is run while searching for the best solution (default=100)}

\item{randomseed}{a random number seed}

\item{iter.max}{the number of iterations used per k-means algorithm (default=1000)}

\item{DirPath}{a directory where the algorithm looks for a previously-saved output (Default=NULL, which turns off this feature)}

\item{Save_Results}{a boolean stating whether to save the output (Default=TRUE)}
}
\value{
Tagged list containing outputs
\describe{
  \item{centers}{a matrix with 2 columns and n_x rows}
  \item{cluster}{A vector with length \code{nrow(loc_orig)} specifying which row of \code{centers} corresponds to each row of loc_orig}
}
}
\description{
\code{Calc_Kmeans} determines the location for a set of knots for approximating spatial variation
}

