% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_utm_to_ll_fn.R
\name{convert_utm_to_ll_fn}
\alias{convert_utm_to_ll_fn}
\title{Convert from UTM to Lat-long}
\usage{
convert_utm_to_ll_fn(eastings, northings, zone, flip_around_dateline = T,
  meters_per_utm_unit = 1000)
}
\arguments{
\item{eastings}{eastings of points, paired to northings}

\item{northings}{northings of points,  paired to eastings}

\item{zone}{UTM zone of points}

\item{flip_around_dateline}{logical stating if points have been flipped around the dateline}

\item{meters_per_utm_unit}{meters per easting/northing}
}
\value{
a data.frame with
\describe{
\item{approx_long}{the approximate transformed longitude of each UTM point}
\item{approx_lat}{the approximate transformed latitude of each UTM point}
}
}
\description{
If eastings and northings have been flipped
around dateline, will convert back and provide estimtes of "original" lat lon
Values should be viewed as estimates for plotting and general pairing
}
\examples{
\dontrun{
eastings <- Spatial_List$loc_x[,'E_km']

northings <- Spatial_List$loc_x[,'N_km']

zone <- Extrapolation_List$zone

flip_around_dateline <- Extrapolation_List$flip_around_dateline

latlongs <- convert_utm_to_ll_fn(eastings = eastings, northings = northings, zone = zone, flip_around_dateline = flip_around_dateline)

ggmap::qmplot(x = approx_long, y = approx_lat, data = latlongs, color = 'red')
}
}
