% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{MapDetails_Fn}
\alias{MapDetails_Fn}
\title{Copy of FishStatsUtils::make_map_info}
\usage{
MapDetails_Fn(...)
}
\description{
Included for continuity when using old scripts
}
\details{
Please use \code{?FishStatsUtils::make_map_info} to see list of arguments and usage
}
