% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{Plot_data_and_knots}
\alias{Plot_data_and_knots}
\title{Copy of FishStatsUtils::plot_data}
\usage{
Plot_data_and_knots(...)
}
\description{
Included for continuity when using old scripts
}
\details{
Please use \code{?FishStatsUtils::plot_data} to see list of arguments and usage
}

