% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{plot_residuals}
\alias{plot_residuals}
\title{Copy of FishStatsUtils::plot_residuals}
\usage{
plot_residuals(...)
}
\description{
Included for continuity when using old scripts
}
\details{
Please use \code{?FishStatsUtils::plot_residuals} to see list of arguments and usage
}

