% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{PlotAniso_Fn}
\alias{PlotAniso_Fn}
\title{Copy of FishStatsUtils::plot_anisotropy}
\usage{
PlotAniso_Fn(...)
}
\description{
Included for continuity when using old scripts
}
\details{
Please use \code{?FishStatsUtils::plot_anisotropy} to see list of arguments and usage
}

