% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{PlotIndex_Fn}
\alias{PlotIndex_Fn}
\title{Copy of FishStatsUtils::plot_biomass_index}
\usage{
PlotIndex_Fn(...)
}
\description{
Included for continuity when using old scripts
}
\details{
Please use \code{?FishStatsUtils::plot_biomass_index} to see list of arguments and usage
}
