% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMap_Fn.R
\name{PlotMap_Fn}
\alias{PlotMap_Fn}
\title{Plot maps with areal results}
\usage{
PlotMap_Fn(MappingDetails, Mat, PlotDF, MapSizeRatio = c(`Width(in)` = 4,
  `Height(in)` = 4), Xlim, Ylim, FileName = paste0(getwd(), "/"), Year_Set,
  Rescale = FALSE, Rotate = 0, Format = "png", Res = 200, zone = NA,
  Cex = 0.01, textmargin = "", add = FALSE, pch = 15,
  outermargintext = c("Eastings", "Northings"), zlim = NULL, Col = NULL,
  Legend = list(use = FALSE, x = c(10, 30), y = c(10, 30)), mfrow = c(1, 1),
  plot_legend_fig = TRUE, land_color = "grey", ignore.na = FALSE, ...)
}
\arguments{
\item{MappingDetails}{tagged list of plot-settings from \code{MapDetails_Fn}}

\item{MapSizeRatio}{Default size for each panel}

\item{Xlim}{xlimits for each panel}

\item{Ylim}{ylimits for each panel}

\item{FileName}{Directory (absolute path) and base for filenames of plots}

\item{Year_Set}{Year names for labeling panels}

\item{Legend}{tagged list specifying insert colorbar
\describe{
  \item{use}{Boolean whether to plot insert colorbar or not}
  \item{x}{Left and right-hand limits for legend in percentage of panel}
  \item{y}{bottom and top limits for legend in percentage of panel}
}}

\item{plot_legend_fig}{Boolean, whether to plot a separate figure for the heatmap legend or not}

\item{land_color}{color for filling in land (use \code{land_color=rgb(0,0,0,alpha=0)} for transparent land)}

\item{...}{arguments passed to \code{par}}
}
\description{
\code{PlotMap_Fn} is a hidden function to plot a map and fill in regions with colors to represent intensity in an areal-interpretion of model results
}
\details{
This function was necessary to build because \code{mapproj::mapproject} as used in \code{maps::map} has difficulties with both rotations (for most projections) and
truncating the cocuntry boundaries within the plotting region (which \code{mapproj::mapproject} appears to do prior to projection,
so that the post-projection is often missing boundaries that are within the plotting rectangle).  I use rectangular projections by default, but Lamberts or Albers conformal
projections would also be useful for many cases.
}

