% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{PlotResultsOnMap_Fn}
\alias{PlotResultsOnMap_Fn}
\title{Copy of FishStatsUtils::plot_maps}
\usage{
PlotResultsOnMap_Fn(...)
}
\description{
Included for continuity when using old scripts
}
\details{
Please use \code{?FishStatsUtils::plot_maps} to see list of arguments and usage
}

