% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Prepare_Extrapolation_Data_Fn.R
\name{Prepare_Extrapolation_Data_Fn}
\alias{Prepare_Extrapolation_Data_Fn}
\title{Build extrapolation grid}
\usage{
Prepare_Extrapolation_Data_Fn(Region, strata.limits, observations_LL = NULL,
  ...)
}
\arguments{
\item{Region}{a character entry that is matched against potential values to determine the region for the extrapolation grid.}

\item{strata.limits}{an input for determining stratification of indices (see example script)}

\item{observations_LL}{a matrix with two columns giving latitude and longitude for each observation (only used when Region doesn't match known entries)}

\item{...}{other objects passed for individual regions (see example script)}
}
\value{
Tagged list used in other functions
\describe{
  \item{a_el}{The area associated with each extrapolation grid cell (rows) and strata (columns)}
  \item{Data_Extrap}{A data frame describing the extrapolation grid}
  \item{zone}{the zone used to convert Lat-Long to UTM by PBSmapping package}
  \item{flip_around_dateline}{a boolean stating whether the Lat-Long is flipped around the dateline during conversion to UTM}
  \item{Area_km2_x}{the area associated with each row of Data_Extrap, in units square-kilometers}
}
}
\description{
\code{Prepare_Extrapolation_Data_Fn} builds an object used to determine areas to extrapolation densities to when calculating indices
}

