% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{QQ_Fn}
\alias{QQ_Fn}
\title{Copy of FishStatsUtils::plot_quantile_diagnostic}
\usage{
QQ_Fn(...)
}
\description{
Included for continuity when using old scripts
}
\details{
Please use \code{?FishStatsUtils::plot_quantile_diagnostic} to see list of arguments and usage
}
