% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Spatial_Information_Fn.R
\name{Spatial_Information_Fn}
\alias{Spatial_Information_Fn}
\title{Build objects related to spatial information}
\usage{
Spatial_Information_Fn(Method = "Grid", Lon, Lat, Extrapolation_List,
  grid_size_km = 50, n_x, ...)
}
\arguments{
\item{Method,}{a character of either "Grid" or "Mesh" where "Grid" is a 2D AR1 process, and "Mesh" is the SPDE method with geometric anisotropy}

\item{Lon,}{Longitude for each sample}

\item{Lat,}{Latitude for each sample}

\item{Extrapolation_List,}{the output from \code{Prepare_Extrapolation_Data_Fn}}

\item{grid_size_km,}{the distance between grid cells for the 2D AR1 grid (determines spatial resolution when Method="Grid")}

\item{n_x,}{the number of vertices in the SPDE mesh (determines the spatial resolution when Method="Mesh")}
}
\value{
Tagged list containing objects for running a VAST model
\describe{
  \item{MeshList}{A tagged list with inputs related to the SPDE mesh}
  \item{GridList}{A tagged list with inputs related to the 2D AR1 grid}
  \item{a_xl}{A data frame with areas for each knot and each strattum}
  \item{loc_UTM}{A data frame with the converted UTM coordinates for each sample}
  \item{Kmeans}{Output from \code{Calc_Kmeans} with knots for a triangulated mesh}
  \item{knot_i}{The knot associated with each sample}
  \item{Method}{The Method input (for archival purposes)}
  \item{loc_x}{The UTM location for each knot}
}
}
\description{
\code{Spatial_Information_Fn} builds a tagged list with all the spatial information needed for \code{Data_Fn}
}

