\name{SpatialDFA-package}
\alias{SpatialDFA-package}
\alias{SpatialDFA}
\docType{package}
\title{
Spatial delta-generalized linear mixed model
}
\description{
Spatial delta-generalized linear mixed model (`SpatialDeltaGLMM`) is used for estimating annual abundance, distribution, and area based on spatially referenced data
}
\details{
\tabular{ll}{
Package: \tab SpatialDeltaGLMM\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2016-09-20\cr
License: \tab GPL-2\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Who wrote it

Maintainer: Who to complain to <yourfault@somewhere.net>
~~ The author and/or maintainer of the package ~~
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
