% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialDeltaGLMM.R
\docType{package}
\name{SpatialDeltaGLMM}
\alias{SpatialDeltaGLMM}
\alias{SpatialDeltaGLMM-package}
\title{Spatial delta-generalized linear mixed model}
\description{
Spatial delta-generalized linear mixed model (`SpatialDeltaGLMM`) is used for estimating annual abundance, distribution, and area based on spatially referenced data
}
\details{
Please see examples folder on github \url{http://github.com/pfmc-assessments/geostatistical_delta-GLMM/tree/master/examples} for up-to-date examples
}
