% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_nwfsc.R
\name{summary_nwfsc}
\alias{summary_nwfsc}
\title{Summarize a VAST model for Northwest Fisheries Science Center}
\usage{
summary_nwfsc(obj, sdreport, savedir = NULL)
}
\arguments{
\item{obj}{Compiled TMB object after estimating MLE of fixed effects}

\item{sdreport}{output from \code{TMB::sdreport}}

\item{savedir}{directory to save tables in CSV format}
}
\value{
Tagged list containing objects for running a VAST model
\describe{
  \item{TableA}{Table of settings}
  \item{TableB}{Table of estimated parameters}
  \item{TableC}{Table of MLE of fixed effects}
}
}
\description{
\code{summary_nwfsc} generates output tables requested by the PFMC SSC
}

