% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_residuals.R
\name{plot_residuals}
\alias{plot_residuals}
\title{Plot Pearson residuals on map}
\usage{
plot_residuals(Lat_i, Lon_i, TmbData, Report, Q, savedir = getwd(),
  MappingDetails, PlotDF, MapSizeRatio = c(`Width(in)` = 4, `Height(in)` = 4),
  Xlim, Ylim, FileName = paste0(getwd(), "/"), Year_Set = NULL,
  Years2Include = NULL, Rescale = FALSE, Rotate = 0, Format = "png",
  Res = 200, zone = NA, Cex = 0.01, add = FALSE, textmargin = NULL,
  pch = NULL, Legend = list(use = FALSE, x = c(10, 30), y = c(10, 30)),
  mfrow = NULL, plot_legend_fig = TRUE, ...)
}
\arguments{
\item{Lat_i}{Latitude for every observation \code{i}}

\item{Lon_i}{Longitude for every observation \code{i}}

\item{TmbData}{a tagged list of data inputs generated by \code{Data_Fn}}

\item{Report}{tagged list of outputs from TMB model via \code{Obj$report()}}

\item{Q}{Output from \code{QQ_Fn}}

\item{MappingDetails}{tagged list of plot-settings from \code{MapDetails_Fn}}

\item{MapSizeRatio}{Default size for each panel}

\item{Xlim}{xlimits for each panel}

\item{Ylim}{ylimits for each panel}

\item{FileName}{Directory (absolute path) and base for filenames of plots}

\item{Year_Set}{Year names for labeling panels}

\item{Years2Include}{integer vector, specifying positions of \code{Year_Set} for plotting (used to avoid plotting years with no data, etc.)}

\item{Legend}{tagged list specifying insert colorbar
\describe{
  \item{use}{Boolean whether to plot insert colorbar or not}
  \item{x}{Left and right-hand limits for legend in percentage of panel}
  \item{y}{bottom and top limits for legend in percentage of panel}
}}

\item{...}{arguments passed to \code{PlotMap_Fn}}

\item{savdir}{directory to use when saving results}
}
\value{
A tagged list of Pearson residuals
\describe{
  \item{Q1_xt}{Matrix of average residuals for encounter/non-encounter component by site \code{x} and year \code{t}}
  \item{Q2_xt}{Matrix of average residuals for positive-catch-rate component by site \code{x} and year \code{t}}
}
}
\description{
\code{plot_residuals} shows average Pearson residual for every knot for encounter probability and positive catch rate components
}
