% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotIndex_Fn.R
\name{PlotIndex_Fn}
\alias{PlotIndex_Fn}
\title{Plot index of abundance}
\usage{
PlotIndex_Fn(TmbData, Sdreport, Year_Set = NULL, Years2Include = NULL,
  DirName = paste0(getwd(), "/"), PlotName = "Index.png",
  interval_width = 1, strata_names = NULL, category_names = NULL,
  use_biascorr = TRUE, plot_legend = TRUE, total_area_km2 = NULL,
  plot_log = FALSE, width = 4, height = 4,
  treat_missing_as_zero = FALSE, create_covariance_table = FALSE, ...)
}
\arguments{
\item{TmbData}{Formatted data inputs, from `SpatialDeltaGLMM::Data_Fn(...)`}

\item{Sdreport}{Standard deviation outputs from TMB model via \code{sdreport(Obj)}}

\item{Year_Set}{Year names for labeling panels}

\item{Years2Include}{integer vector, specifying positions of \code{Year_Set} for plotting (used to avoid plotting years with no data, etc.)}

\item{DirName}{Directory for saving plot and table}

\item{PlotName}{Name for plot}

\item{interval_width}{width for confidence intervals}

\item{strata_names}{names for spatial strata}

\item{category_names}{names for categories (if using package `VAST`)}

\item{use_biascorr}{Boolean, whether to use bias-corrected estimates if available}

\item{plot_legend}{Add legend for labelling colors}

\item{total_area_km2}{Total area for calculating a design-based estimator using one design-stratum (only recommended for model exploration)}

\item{plot_log}{Boolean, whether to plot y-axis in log-scale}

\item{width}{plot width in inches}

\item{height}{plot height in inches}

\item{treat_missing_as_zero}{Boolean whether to treat years and species with no (or only NA) data as instances where the index should be zero}

\item{...}{Other inputs to `par()`}
}
\value{
Return Tagged list of output
\describe{
  \item{Table}{table of index estimates by stratum and year, e.g., for including in an assessment model}
}
}
\description{
\code{PlotIndex_Fn} plots an index proportion to population abundance
}

