% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotResultsOnMap_Fn.r
\name{PlotResultsOnMap_Fn}
\alias{PlotResultsOnMap_Fn}
\title{Plot standard maps}
\usage{
PlotResultsOnMap_Fn(plot_set = 3, MappingDetails, Report, Sdreport = NULL,
  Nknots = Inf, PlotDF, MapSizeRatio = c(`Width(in)` = 4, `Height(in)` = 4),
  Xlim, Ylim, FileName = paste0(getwd(), "/"), Year_Set = NULL,
  Years2Include = NULL, Rescale = FALSE, Rotate = 0, Format = "png",
  Res = 200, zone = NA, Cex = 0.01, add = FALSE,
  category_names = NULL, textmargin = NULL, pch = NULL,
  Legend = list(use = FALSE, x = c(10, 30), y = c(10, 30)), mfrow = NULL,
  plot_legend_fig = TRUE, ...)
}
\arguments{
\item{plot_set}{integer-vector defining plots to create
\describe{
  \item{plot_set=1}{Probability of encounter/non-encounter}
  \item{plot_set=2}{Log-expected positive catch rate}
  \item{plot_set=3}{Log-predicted density (product of encounter probability and positive catch rates)}
  \item{plot_set=4}{Log-positive catch rates (rescaled)}
  \item{plot_set=5}{Log-predicted density (rescaled)}
  \item{plot_set=6}{Spatio-temporal variation in encounter probability}
  \item{plot_set=7}{Spatio-temporal variation in log-positive catch rates}
  \item{plot_set=8}{Linear predictor for encounter probability}
  \item{plot_set=9}{Linear predictor for positive catch rates}
  \item{plot_set=10}{Coefficient of variation for predicted density (available only if \code{Data_Fn(...,Options=c('SD_site_logdensity'=1,...))}}
}}

\item{MappingDetails}{tagged list of plot-settings from \code{MapDetails_Fn}}

\item{Report}{tagged list of outputs from TMB model via \code{Obj$report()}}

\item{Sdreport}{Standard deviation outputs from TMB model via \code{sdreport(Obj)}}

\item{Nknots}{Number of knots used for plotting (default is all grid cells in \code{Extrapolation_grid}}

\item{MapSizeRatio}{Default size for each panel}

\item{Xlim}{xlimits for each panel}

\item{Ylim}{ylimits for each panel}

\item{FileName}{Directory (absolute path) and base for filenames of plots}

\item{Year_Set}{Year names for labeling panels}

\item{Years2Include}{integer vector, specifying positions of \code{Year_Set} for plotting (used to avoid plotting years with no data, etc.)}

\item{category_names}{character vector specifying names for different categories (only used for R package \code{VAST})}

\item{Legend}{tagged list specifying insert colorbar
\describe{
  \item{use}{Boolean whether to plot insert colorbar or not}
  \item{x}{Left and right-hand limits for legend in percentage of panel}
  \item{y}{bottom and top limits for legend in percentage of panel}
}}

\item{...}{arguments passed to \code{PlotMap_Fn}}
}
\value{
Mat_xt a matrix (rows: modeled knots; column: modeled year) for plotted output of last element of \code{plot_set}
}
\description{
\code{PlotResultsOnMap_Fn} plots a standard set of diagnostic maps
}

