% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{Prepare_Extrapolation_Data_Fn}
\alias{Prepare_Extrapolation_Data_Fn}
\title{Copy of FishStatsUtils::make_extrapolation_data}
\usage{
Prepare_Extrapolation_Data_Fn(...)
}
\description{
Included for continuity when using old scripts
}
\details{
Please use \code{?FishStatsUtils::Prepare_Extrapolation_Data_Fn} to see list of arguments and usage
}

