\name{SpatialDeltaGLMM-package}
\alias{SpatialDeltaGLMM-package}
\alias{SpatialDeltaGLMM}
\docType{package}
\title{ 
Package for conducting spatial estimation of CPUE index standardization (fisheries, ecology, etc).
}
\description{
SpatialDeltaGLMM: 

(1) Is an R package for implementing a spatial delta-generalized linear mixed model (delta-GLMM) for use when standardizing fishery-independent index data for U.S. West Coast surveys with 

(2) built in diagnostic functions and model-comparison tools, and 

(3) is intended to improve analysis speed, replicability, peer-review, and interpretation of index standardization methods
}
\details{
\tabular{ll}{
Package: \tab tmp\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-10-25\cr
License: \tab What license is it under?\cr
}}
\author{
James Thorson, 

Maintainer: James Thorson, James.Thorson@noaa.gov
}
\references{
Please see https://github.com/nwfsc-assess/geostatistical_delta-GLMM for an up-to-date list of references
}
\keyword{ package }
\seealso{
See \code{\link{Data_Fn}} for an explanation of data inputs
 
See \code{\link{Param_Fn}} for an explanation of coefficients (fixed and random effects) estimated by TMB 
}
\examples{
Please see https://github.com/nwfsc-assess/geostatistical_delta-GLMM/tree/master/examples for examples of how to prepare data, run the model, and process results
}
